[4/n] - Move external models to their own file
