+++ src/dragoneye/__init__.py
@@ -1,10 +1,12 @@
 from .classification import (
     Classification,
+)
+from .client import Dragoneye
+from .models import (
     ClassificationObjectPrediction,
     ClassificationPredictImageResponse,
     ClassificationTraitRootPrediction,
 )
-from .client import Dragoneye
 from .types.common import NormalizedBbox, TaxonID, TaxonPrediction, TaxonType
 from .types.media import Media
